+++ plpnfsd/rfsv_api.h
@@ -17,7 +17,7 @@ extern long rfsv_rmdir(const char *name);
 extern long rfsv_remove(const char *name);
 extern long rfsv_rename(const char *oldname, const char *newname);
 extern long rfsv_fclose(long handle);
-extern long rfsv_fcreate(long attr, const char *name, long *handle);
+extern long rfsv_fcreate(long attr, const char *name, u_int32_t *handle);
 extern long rfsv_read(char *buf, long offset, long len, char *name);
 extern long rfsv_write(char *buf, long offset, long len, char *name);
 extern long rfsv_getattr(const char *name, long *attr, long *size, long *time);
